Created config loaders and changed Config object accordingly.
The following loaders exist:
- ConfigLoader (base class)
- ConfigLoaderFileSystem (load from file system, receives a folder)
- ConfigLoaderMultiplexor (load from one of multiple loaders, receives
  a list of loaders)
- ConfigLoaderStrings (loads from a dictionary of strings)

Also, the function getDefaultLoader was created. This function returns a
multiplexor with two file system loaders (one for "./config" and the other
for "/etc/ecommerce").

All loaders accept a list of fragments (meaning: configuration fragments).
The default list of fragments has: "global" and "local".

Modified the tests to match the new architecture.