Added keychain functionality to Config object
The keychain has a "fetch" method that retrieves the named key if
the passed in value is of the form "keychain:{{keyname}}" or returns
the same passed in value (this allows having passwords straith in
the config). For **SECURITY** the keychain is a readonly property.

The keychain requires the following configuration at the top level:

    "keychain" : {
        "file" : "keychain.yaml",
        "dirs" : [ "./config", "/etc/ecommerce" ]
    }

The file name and directories can obviously change. The keychain file
itself has a sintax like the following:

    master-key:
        # some hexa value
        value:      44335522661177
        algorithm:  master-key
    master-db:
        value:      somekey
        algorithm:  # no algorithm => clear text
    some-other-key:
        # some hexa value
        value:      44335522661177
        algorithm:  3DES-cbc
        key:        master-key

Support for encryption is not there yet, but development can proceeed as
if it were.

Added a set of tests for the keychain.