Modified the keychain to be a more generic key data repository.
The KeyChain file now has a "data" entry for each key. That data
entry is a hashmap that can store multiple values. For example
"password" (for a database), "id" and "key" (for AWS), etc. Examples
of this are in the samples/keychain.yaml file.

Access to each individual key is now on the form:

       keychain:{{keyname}}:{{keyvalue}}