Added the possibility to indicate that a database has loose types (SQLite3) and a method to get the flag.
Added new method hasLooseTypes(dbname) to ecommerce.db. It returns
True if the database has loosetypes (SQLite3) or false if it has
strict types. This means: is True if a column can have a value that
can vary from row to row.

Changed samples/global.yaml to reflect the new configuration property.