Fixes the problem with filenames longer than 260 characters for products.
What it does is calculate the relative "url" part of the name and then if it
is longer than 200 chars it is truncated. Then the "--<id product>" is appended.

This gives room for a file extension (.htm, .js, etc) and for a prefix configured
directory (such as a drive and folder, ex: s:/tematika)